+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## owncloud-onlyoffice (1.0.1)
 - fix exception when versions app is disabled
+- adding protocol to document server url
 
 
 ## owncloud-onlyoffice (1.0.0)